Renamed files to final name
