+++ frontend/src/App.js
@@ -3,7 +3,7 @@ import './App.css'
 import AppBar from '@material-ui/core/AppBar'
 import Toolbar from '@material-ui/core/Toolbar'
 import Typography from '@material-ui/core/Typography'
-import { ToDoLists } from './todos/components/ToDoLists'
+import { ToDoLists } from './todos/components/TodoLists'
 
 const MainAppBar = () => {
   return <AppBar position='static' color='primary'>
