+++ frontend/src/todos/components/TodoLists.jsx
@@ -7,7 +7,7 @@ import ListItemText from '@material-ui/core/ListItemText'
 import ListItemIcon from '@material-ui/core/ListItemIcon'
 import ReceiptIcon from '@material-ui/icons/Receipt'
 import Typography from '@material-ui/core/Typography'
-import { ToDoListForm } from './ToDoListForm'
+import { ToDoListForm } from './TodoListForm'
 
 const sleep = ms => new Promise(resolve => setTimeout(resolve, ms))
